Crate rel_core

rel_core is a set of portable, relative-pointer based replacements for core library types.

Re-exports

pub use self::rel_ptr::RelPtr;

pub use self::rel_ref::RelRef;

Modules

export

Public re-exports of dependencies.

option

A value that may or may not exist.

rel_mem

Basic functions for dealing with relative values in memory.

rel_ptr

Relative pointers and related types.

rel_ref

Relative pointers and related types.

rel_tuple

Relative versions of tuples.

Structs

Char

A portable char.

DefaultBasis

The default Basis.

F32

A portable f32.

F64

A portable f64.

I16

A portable i16.

I32

A portable i32.

I64

A portable i64.

I128

A portable i128.

U16

A portable u16.

U32

A portable u32.

U64

A portable u64.

U128

A portable u128.

Traits

Basis

A selection of types to use in place of isize and usize.

BasisPointee

A Pointee with metadata that changes with a choice of basis.

Emplace

A value emplacer.

EmplaceExt

An extension trait for Emplace that provides a variety of convenient emplacement methods.

Move

An emplaced value that can be moved.

MoveExt

An extension trait for Move that provides a variety of convenient movement methods.

Portable

A type that has the same representation on all targets.

Type Definitions

Bool

Alias for bool.

I8

Alias for i8.

U8

Alias for u8.

Unit

Alias for ().

Derive Macros

Move

Derives Move on the annotated type.

Portable

Derives Portable on the annotated type.