Trait rel_core::Portable

pub unsafe trait Portable { }

A type that has the same representation on all targets.

Safety

Portable types must have the same layout and bytewise representation on all targets.